backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
SQLAlchemy==2.0.36
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
email-validator==2.2.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
authlib==1.3.1
httpx==0.27.2
python-dotenv==1.0.1
boto3==1.34.158
pywebpush==1.14.0
jsonschema==4.23.0
starlette==0.40.0
redis[hiredis]==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3
stripe==7.7.0
